security

+++ admin/includes/comment_perpost.php
@@ -2,7 +2,7 @@
 
 if (isset($_POST['checkBoxesArray'])) {
     foreach ($_POST['checkBoxesArray'] as $checkBoxesValues) {
-        $bulk_option = $_POST['bulk_option'];
+        $bulk_option = escape($_POST['bulk_option']);
 
         switch ($bulk_option) {
             case 'approved':
@@ -58,18 +58,18 @@ if (isset($_POST['checkBoxesArray'])) {
             <tbody>
                 <?php
                 if (isset($_GET['p_id'])) {
-                    $post_id = $_GET['p_id'];
+                    $post_id = escape($_GET['p_id']);
                     $query = "SELECT * FROM comments WHERE comment_post_id = $post_id";
                     $comment_posts = mysqli_query($connection, $query);
 
                     while ($row = mysqli_fetch_assoc($comment_posts)) {
-                        $comment_id = $row['comment_id'];
-                        $comment_author = $row['comment_author'];
-                        $comment_content = $row['comment_content'];
-                        $comment_email = $row['comment_email'];
-                        $comment_status = $row['comment_status'];
-                        $post_comment = $row['comment_post_id'];
-                        $comment_date = $row['comment_date'];
+                        $comment_id =  escape($row['comment_id']);
+                        $comment_author =  escape($row['comment_author']);
+                        $comment_content =  escape($row['comment_content']);
+                        $comment_email =  escape($row['comment_email']);
+                        $comment_status =  escape($row['comment_status']);
+                        $post_comment =  escape($row['comment_post_id']);
+                        $comment_date =  escape($row['comment_date']);
 
 
 
@@ -104,30 +104,63 @@ if (isset($_POST['checkBoxesArray'])) {
                 <?php
                     //approve comments
                     if (isset($_GET['approve'])) {
-                        $comment_id_update = $_GET['approve'];
+                        $comment_id_update = escape($_GET['approve']);
 
-                        $query = "UPDATE comments SET comment_status = 'approved' WHERE comment_id = $comment_id_update ";
-                        $approve_query = mysqli_query($connection, $query);
-                        header("location: ./comments.php?source=c_perpost&p_id=$post_id");
+                        $user_id = escape($_SESSION['user_id']);
+                        $user_id = mysqli_real_escape_string($connection, $_SESSION['user_id']);
+                        if (isset($user_id)) {
+                            $query = "SELECT role FROM users WHERE user_id = $user_id";
+                            $get_role = mysqli_query($connection, $query);
+                            $row = mysqli_fetch_assoc($get_role);
+                            $role = $row['role'];
+
+                            if ($role === "admin") {
+                                $query = "UPDATE comments SET comment_status = 'approved' WHERE comment_id = $comment_id_update ";
+                                $approve_query = mysqli_query($connection, $query);
+                                header("location: ./comments.php?source=c_perpost&p_id=$post_id");
+                            }
+                        }
                     }
 
                     //unapprove comments
                     if (isset($_GET['unapprove'])) {
-                        $comment_id_update = $_GET['unapprove'];
+                        $comment_id_update = escape($_GET['unapprove']);
 
-                        $query = "UPDATE comments SET comment_status = 'unapproved' WHERE comment_id = $comment_id_update ";
-                        $unapprove_query = mysqli_query($connection, $query);
-                        header("location: ./comments.php?source=c_perpost&p_id=$post_id");
+                        $user_id = escape($_SESSION['user_id']);
+                        $user_id = mysqli_real_escape_string($connection, $_SESSION['user_id']);
+                        if (isset($user_id)) {
+                            $query = "SELECT role FROM users WHERE user_id = $user_id";
+                            $get_role = mysqli_query($connection, $query);
+                            $row = mysqli_fetch_assoc($get_role);
+                            $role = $row['role'];
+
+                            if ($role === "admin") {
+                                $query = "UPDATE comments SET comment_status = 'unapproved' WHERE comment_id = $comment_id_update ";
+                                $unapprove_query = mysqli_query($connection, $query);
+                                header("location: ./comments.php?source=c_perpost&p_id=$post_id");
+                            }
+                        }
                     }
 
 
                     //delete comment
                     if (isset($_GET['delete'])) {
-                        $comment_id_delete = $_GET['delete'];
+                        $comment_id_delete = escape($_GET['delete']);
 
-                        $query = "DELETE FROM comments WHERE comment_id = $comment_id_delete ";
-                        $delete_query = mysqli_query($connection, $query);
-                        header("location: ./comments.php?source=c_perpost&p_id=$post_id");
+                        $user_id = escape($_SESSION['user_id']);
+                        $user_id = mysqli_real_escape_string($connection, $_SESSION['user_id']);
+                        if (isset($user_id)) {
+                            $query = "SELECT role FROM users WHERE user_id = $user_id";
+                            $get_role = mysqli_query($connection, $query);
+                            $row = mysqli_fetch_assoc($get_role);
+                            $role = $row['role'];
+
+                            if ($role === "admin") {
+                                $query = "DELETE FROM comments WHERE comment_id = $comment_id_delete ";
+                                $delete_query = mysqli_query($connection, $query);
+                                header("location: ./comments.php?source=c_perpost&p_id=$post_id");
+                            }
+                        }
                     }
                 }
